+++ executor-service/src/main/java/executor/service/exception/logstorage/ConnectionFailedException.java
@@ -1,4 +1,4 @@
-package executor.service.exception.dbmanagerexception;
+package executor.service.exception.logstorage;
 
 public class ConnectionFailedException extends RuntimeException {
     public ConnectionFailedException(Throwable cause) {
